# server/docs/swagger.yaml
basePath: /
...
info:
  contact: {}
  description: 若依 Go 版本接口文档
  title: RuoYi-Go API
  version: "1.0"
paths:
...
  /system/dict/data:
    post:
      consumes:
      - application/json
      parameters:
      - description: 字典数据信息
        in: body
        name: body
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 操作结果
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 新增字典数据
      tags:
      - 字典数据
    put:
      consumes:
      - application/json
      parameters:
      - description: 字典数据信息
        in: body
        name: body
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 操作结果
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 修改字典数据
      tags:
      - 字典数据
  /system/dict/data/{id}:
    get:
      parameters:
      - description: 字典数据ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 字典数据信息
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 查询字典数据详细
      tags:
      - 字典数据
  /system/dict/data/{ids}:
    delete:
      parameters:
      - description: 字典数据ID，多个用逗号分隔
        in: path
        name: ids
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 操作结果
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 删除字典数据
      tags:
      - 字典数据
  /system/dict/data/export:
    post:
      produces:
      - application/octet-stream
      responses:
        "200":
          description: Excel文件
          schema:
            type: file
      security:
      - BearerAuth: []
      summary: 导出字典数据
      tags:
      - 字典数据
  /system/dict/data/list:
    get:
      parameters:
      - description: 字典类型
        in: query
        name: dictType
        type: string
      - description: 字典标签
        in: query
        name: dictLabel
        type: string
      - description: 状态
        in: query
        name: status
        type: string
      - description: 页码
        in: query
        name: pageNum
        type: integer
      - description: 每页数量
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 字典数据列表
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 查询字典数据列表
      tags:
      - 字典数据
  /system/dict/data/types:
    post:
      consumes:
      - application/json
      parameters:
      - description: 字典类型列表
        in: body
        name: body
        required: true
        schema:
          items:
            type: string
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: 字典数据
          schema:
            additionalProperties: true
            type: object
      summary: 根据字典类型查询字典数据信息
      tags:
      - 字典数据
...
securityDefinitions:
  BearerAuth:
    description: Bearer {token}
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"